+++ sample-project/bin/prod.dart
@@ -1,4 +1,0 @@
-import 'package:sample_project/sample_project.dart';
-import 'package:angel_production/angel_production.dart';
-
-main(List<String> args) => Runner('angel', configureServer).run(args);