+++ sample-project/lib/src/config/plugins/plugins.dart
@@ -1,8 +1,0 @@
-library sample_project.src.config.plugins;
-
-import 'dart:async';
-import 'package:angel_framework/angel_framework.dart';
-
-Future configureServer(Angel app) async {
-  // Include any plugins you have made here.
-}